refactored imports to use index.js
