+++ src/App.js
@@ -1,10 +1,6 @@
 import { useState } from 'react';
-import Header from './components/Header';
-import Footer from './components/Footer';
-import About from './pages/About';
-import Projects from './pages/Projects';
-import Contact from './pages/Contact';
-import Resume from './pages/Resume';
+import { Header, Footer } from './components';
+import { About, Projects, Contact, Resume } from './pages';
 import './App.css';
 
 // needed to override bootstrap
