+++ src/pages/Projects.js
@@ -1,4 +1,4 @@
-import Project from '../components/ProjectCard';
+import { ProjectCard } from '../components';
 // array of projects with keys used in .map()
 import projects from '../assets/projectList.js';
 
@@ -10,7 +10,7 @@ function Projects() {
       <ol className="d-flex flex-wrap justify-content-around">
         {projects.map(({ title, githubLink, deployedLink, img, tech }) => {
           return (
-            <Project
+            <ProjectCard
               key={title}
               project={{ title, githubLink, deployedLink, img, tech }}
             />
